// src/types/partials/TestScriptTest.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TTestScriptAction1 } from '../partials/TestScriptAction1';

export type TTestScriptTest = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    name?: String;
    description?: String;
    action: TTestScriptAction1[];
};